+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### v2.0.0 (2020/4/16)
+
+- **Breaking**: All values are now either numbers or `null`. There are no strings containing numbers, and no `"N/A" values.
+- **New feature**: A TypeScript definition file is now exported from this library
+
 ### v1.1.1 (2020/4/16)
 
 - **Bugfix**: Fixed `cape` value